refactor: move search function to fetch kopps results closesr to components
